+++ Core/Contracts/Response/HarvestNodeUnlockResponse.cs
@@ -1,10 +1,11 @@
 ﻿using IdelPog.Core.Contracts.Enum;
 
-namespace IdelPog.Progression.Contracts
+namespace IdelPog.Core.Contracts.Response
 {
-    public readonly record struct HarvestNodeUnlock
+    public readonly record struct HarvestNodeUnlockResponse
     {
         public required SkillID SkillID { get; init; }
+        public required ItemID ItemID { get; init; }
         public required byte SkillLevel { get; init; }
     }
 }
